robust play stop and clean up

+++ code/bot/cogs/music_remaster.py
@@ -19,12 +19,16 @@ class MusicCog(commands.Cog):
         self.player = None
         # Disconnect timer task
         self.disconnect_timer = None
+        # Store the voice channel id for reconnection purposes
+        self.current_voice_channel_id = None
+        # Lock to secure the UI player
+        self.ui_lock = asyncio.Lock()
 
     # Creating the group play, which is the master command to play a music
     play_group = app_commands.Group(name='play', description='Play a music from available music sources.')
     
     # Sub-command to play a music from a Youtube link or title
-    @play_group.command(name='youtube', description='Play a music from a Youtube link or title.')
+    @play_group.command(name='youtube', description='⏯️ Play a music from a Youtube link or title.')
     @app_commands.describe(query="Link or URL")
     @app_commands.autocomplete(query=YTDownload.preview_results)
     async def youtube(self, interaction: discord.Interaction, query : str):
@@ -110,20 +114,68 @@ class MusicCog(commands.Cog):
                                 color=discord.Color.red()
                             )
             await interaction.followup.send(embed=error_embed, ephemeral=True)
+    
+    # Command to stop the current music and clear the waiting list
+    @play_group.command(name='stop', description='⏹️ Stop the current music and clear the waiting list.')
+    async def stop(self, interaction: discord.Interaction):
+        # Get bot's voice_client
+        voice_client = interaction.guild.voice_client
+        # Ensure the command can be ran
+        try:
+            await self.ensure_context(interaction)
+        except:
+            # Return if we could not ensure
+            return
+        # Call the cleanup function that will ensure variables cleanup
+        await self.cleanup()
+        # Try to stop the voice
+        if voice_client and voice_client.is_playing():
+            try:
+                voice_client.stop()
+            except Exception as e:
+                print(f'[BOT] --- Voice client stop error : {e}')
+        stopped_embed = discord.Embed(
+                                title=f'⏹️ Music stopped by -{interaction.user.name}-.',
+                                description='Leaving the voice channel in 1 minute.',
+                                color=discord.Color.green()
+                            )
+        await interaction.response.send_message(embed=stopped_embed)
+        # Start the disconnect timer
+        self.disconnect_timer = asyncio.create_task(self.disconnect_after_delay(interaction))
+
+    # Cleaning up the bot's variables
+    async def cleanup(self):
+        print("[BOT] --- Cleaning up...")
+        # Delete previous player if it exists
+        async with self.ui_lock:
+            if self.player:
+                try:
+                    await self.player.delete()
+                except discord.NotFound:
+                    # Message was already deleted (by user or by ensure_player)
+                    pass
+        # Clear variables
+        self.waitlist.clear()
+        self.player = None
+        self.current_voice_channel_id = None
+        if self.disconnect_timer:
+            self.disconnect_timer.cancel()
+            self.disconnect_timer = None
 
     # Function to ensure the player message is always at the bottom of the chat
     async def ensure_player(self):
-        try:
-            # Retrieve the original player message
-            msg = self.player.embeds[0]
-            # Retrieve the original player message channel
-            channel = self.player.channel
-            # Delete the old player message
-            await self.player.delete()
-            # Send it again
-            self.player = await channel.send(embed=msg)
-        except discord.NotFound:
-            pass
+        async with self.ui_lock:
+            try:
+                # Retrieve the original player message
+                msg = self.player.embeds[0]
+                # Retrieve the original player message channel
+                channel = self.player.channel
+                # Delete the old player message
+                await self.player.delete()
+                # Send it again
+                self.player = await channel.send(embed=msg)
+            except discord.NotFound:
+                pass
     
     async def disconnect_after_delay(self, interaction: discord.Interaction):
         print("[BOT] --- Started inactivity timer.")
@@ -133,26 +185,104 @@ class MusicCog(commands.Cog):
         # Get the bot's voice client
         voice_client = interaction.guild.voice_client
 
-        # If the voice client is still
+        # Reset timer variable
+        self.disconnect_timer = None
+
+        # If the voice client is still up
         if voice_client and voice_client.is_connected():
-            # Disconnect it
             await voice_client.disconnect()
-            # Reset bot's variables
-            self.player = None
-            self.disconnect_timer = None
             print("[BOT] --- Disconnected for inactivity.")
+
+    # Ensure the user can actually execute this command (i.e the user and the bot are in the same voice channel % other things)
+    async def ensure_context(self, interaction: discord.Interaction):
+        # Get the bot's voice client
+        bot_voice = interaction.guild.voice_client
+        # Get the user's voice
+        user_voice = interaction.user.voice
+        # If the user is not in a voice channel
+        if not user_voice:
+            no_user_embed = discord.Embed(
+                    title='❌ You are not in a voice channel.',
+                    color=discord.Color.red()
+                )
+            await interaction.response.send_message(embed=no_user_embed, ephemeral=True, delete_after=10)
+            raise commands.CommandError("User not in a voice channel.")
+        
+        # If the bot is not in a voice channel or not playing
+        if not bot_voice or not bot_voice.is_playing():
+            no_user_embed = discord.Embed(
+                    title='❌ The bot is not playing.',
+                    color=discord.Color.red()
+                )
+            await interaction.response.send_message(embed=no_user_embed, ephemeral=True, delete_after=10)
+            raise commands.CommandError("Bot not in a voice channel.")
+
+        # If the user is not in the bot's voice channel
+        if user_voice.channel != bot_voice.channel:
+            no_user_embed = discord.Embed(
+                    title='❌ You are not in the bot\'s voice channel.',
+                    color=discord.Color.red()
+                )
+            await interaction.response.send_message(embed=no_user_embed, ephemeral=True, delete_after=10)
+            raise commands.CommandError("UUser and bot not in the same voice channel.")
+        
+
+    # Ensure the bot is still in the original voice channel to play the next song
+    async def ensure_next(self, interaction: discord.Interaction):
+        # Get the bot's voice
+        voice_client = interaction.guild.voice_client
+
+        # If no original channel id is stored, it means that the bot has been manually stopped, so cleanup has already been done
+        if not self.current_voice_channel_id:
+            return None
+
+        # Normal case, everything if fine
+        if voice_client and voice_client.is_connected():
+            return voice_client
+
+        # At this point, the bot is not connected anymore and this is not normal
+        print("[BOT] --- Bot disconnected but waiting list not empty, attempting to reconnect...")
+        try:
+            # Get the original voice channel
+            channel = self.bot.get_channel(self.current_voice_channel_id)
+            # Check if channel exists and there is at least one person inside
+            if channel and len(channel.members) > 0:
+                # Try to connect to it
+                await channel.connect()
+                print("[BOT] --- Successfully reconnected.")
+                # Return the new voice_client
+                return interaction.guild.voice_client
+            else:
+                print("[BOT] --- Channel does not exist or no one inside.")
+        except Exception as e:
+            print(f"[BOT] --- Failed to reconnect to channel : {e}")
+        
+        # If we're still there, cleanup and tell the user that the reconnect failed
+        await self.cleanup()
+        no_voice_embed = discord.Embed(
+                    title=f'❌ Can\'t reconnect and keep playing in the original voice channel.',
+                    description="If you wish to play again, join a voice channel and call /play.",
+                    color=discord.Color.red()
+                )
+        await interaction.channel.send(embed=no_voice_embed)
+        return None
 
     # Function to play a song in the bot's voice channel, executed everytime a song ends if the waiting list has something in it
     async def play_next(self, interaction: discord.Interaction):
-        # Get the bot's voice
-        voice_client = interaction.guild.voice_client
+        # Ensure we can play the next song with a valid voice_client
+        voice_client = await self.ensure_next(interaction)
+        # If voice_client not valid, we cannot continue, cleanup has been done and user has been notified
+        if not voice_client:
+            return
         # Delete previous player if it exists
-        if self.player:
-            try:
-                await self.player.delete()
-            except discord.NotFound:
-                # Message was already deleted (by user or by ensure_player)
-                pass
+        async with self.ui_lock:
+            if self.player:
+                try:
+                    await self.player.delete()
+                except discord.NotFound:
+                    # Message was already deleted (by user or by ensure_player)
+                    pass
+                self.player = None
         # If something is in the waiting list
         if len(self.waitlist) > 0:
             # Retrieve the data from the next song in queue
@@ -170,7 +300,8 @@ class MusicCog(commands.Cog):
             playing_embed.add_field(name='Likes', value=data['like_count'], inline=True)
             playing_embed.add_field(name='Upload date', value=data['upload_date'], inline=True)
             playing_embed.set_footer(text=f'URL: {data["webpage_url"]}')
-            self.player = await interaction.channel.send(embed=playing_embed)
+            async with self.ui_lock:
+                self.player = await interaction.channel.send(embed=playing_embed)
 
             # Audio source
             source = discord.FFmpegPCMAudio(f'code/bot/cogs/temp_songs/{data["id"]}.{data["ext"]}')
@@ -184,9 +315,12 @@ class MusicCog(commands.Cog):
                 title='🏁 Finished playing '+os.getenv('FINISHED_PLAYING'),
                 color=discord.Color.green()
             )
-            self.player = await interaction.channel.send(embed=finished_embed)
-            # Start the disconnect timer
-            self.disconnect_timer = asyncio.create_task(self.disconnect_after_delay(interaction))
+            await interaction.channel.send(embed=finished_embed, delete_after=20)
+            # Start the disconnect timer if the bot is still in the voice channel
+            if voice_client and voice_client.is_connected():
+                # Clean up before the disconnect happens
+                await self.cleanup()
+                self.disconnect_timer = asyncio.create_task(self.disconnect_after_delay(interaction))
         
     # Ensure the bot is/will connected/connect to the user voice channel
     async def ensure_voice(self, interaction: discord.Interaction):
@@ -238,6 +372,10 @@ class MusicCog(commands.Cog):
             else:
                 await interaction.response.send_message(embed=error_embed, ephemeral=True, delete_after=10)
             raise error_exception
+        
+        # Store the voice channel after ensuring connection
+        if interaction.guild.voice_client:
+            self.current_voice_channel_id = interaction.guild.voice_client.channel.id
         
 # This function is used to setup the cog
 async def setup(bot):
